Add `import "path"` binds and exps
An `import "path"` expression "statically" loads the module implemented at
"path".  The "path" may either name a specific `.1ml` (and the suffix may be
omitted) file or a directory containing an `index.1ml` file.  Additional there
may be an associated `.1mls` file containing a signature.  If the "path" begins
with "./" or "../" it is interpreted as a relative path and only attempted in
the directory of the file from which the `import "path"` was made.  Likewise an
absolute path is only interpreted as an absolute path.  Otherwise a set of
module load directories are searched for a matching module.

`import` expressions are effectively hoisted out of their containing module
expression and the referenced files are then evaluated once and their results
are bound to variables.  An `import "path"` expression is then interpreted as a
reference to a variable.

An `import "path"` binding is sugar for `...import "path"`.